jarvis_server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-multipart>=0.0.12
edge-tts>=6.1.10
openai-whisper>=20240930
